server/smart_cast_db/requirements.txt
psycopg[binary]>=3.2,<4
python-dotenv>=1.0,<2
